PLAT-696: better vstatereport validation
